move: enum type 위치 변경
