--- src/main/java/seong/onlinestudy/domain/PostCategory.java
@@ -1,5 +1,0 @@
-package seong.onlinestudy.domain;
-
-public enum PostCategory {
-    INFO, QUESTION, CHAT
-}
