+++ src/main/java/seong/onlinestudy/enumtype/GroupCategory.java
@@ -1,4 +1,4 @@
-package seong.onlinestudy.domain;
+package seong.onlinestudy.enumtype;
 
 public enum GroupCategory {
     JOB, GOV, UNIV, HIGH, BOOK, LANG, CERT, IT,  ETC
