+++ src/main/java/seong/onlinestudy/enumtype/OrderBy.java
@@ -1,4 +1,4 @@
-package seong.onlinestudy.request;
+package seong.onlinestudy.enumtype;
 
 public enum OrderBy {
     CREATEDAT, MEMBERS, TIME, ATTENDANCE
