+++ src/test/java/seong/onlinestudy/controller/TicketControllerTest.java
@@ -14,10 +14,10 @@ import org.springframework.restdocs.payload.JsonFieldType;
 import org.springframework.test.util.ReflectionTestUtils;
 import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.ResultActions;
-import seong.onlinestudy.docs.DocumentFormatGenerator;
 import seong.onlinestudy.domain.*;
 import seong.onlinestudy.dto.MemberTicketDto;
 import seong.onlinestudy.dto.TicketDto;
+import seong.onlinestudy.enumtype.TicketStatus;
 import seong.onlinestudy.request.ticket.TicketCreateRequest;
 import seong.onlinestudy.request.ticket.TicketGetRequest;
 import seong.onlinestudy.service.TicketService;
