+++ src/test/java/seong/onlinestudy/repository/MemberRepositoryTest.java
@@ -7,6 +7,7 @@ import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import seong.onlinestudy.domain.*;
+import seong.onlinestudy.enumtype.GroupRole;
 
 import java.time.LocalDateTime;
 import java.util.ArrayList;
