+++ src/test/java/seong/onlinestudy/service/GroupServiceTest.java
@@ -14,6 +14,7 @@ import seong.onlinestudy.MyUtils;
 import seong.onlinestudy.domain.*;
 import seong.onlinestudy.dto.GroupDto;
 import seong.onlinestudy.dto.GroupStudyDto;
+import seong.onlinestudy.enumtype.GroupRole;
 import seong.onlinestudy.exception.PermissionControlException;
 import seong.onlinestudy.repository.GroupMemberRepository;
 import seong.onlinestudy.repository.GroupRepository;
